Crate srml_aura

Aura Module

aura::Trait

Module

Overview

The Aura module extends Aura consensus by managing offline reporting.

Interface

Public Functions

slot_duration - Determine the Aura slot-duration based on the Timestamp module configuration.

Related Modules

Staking: The Staking module is called in Aura to enforce slashing if validators miss a certain number of slots (see the StakingSlasher struct and associated method).

Timestamp: The Timestamp module is used in Aura to track consensus rounds (via slots).

Consensus: The Consensus module does not relate directly to Aura, but serves to manage offline reporting by implementing ProvideInherent in a similar way.

References

If you're interested in hacking on this module, it is useful to understand the interaction with substrate/core/inherents/src/lib.rs and, specifically, the required implementation of ProvideInherent and ProvideInherentData to create and check inherents.

Re-exports

pub use timestamp;

Structs

AuraReport
A report of skipped authorities in Aura.

Authorities
The current authorities

GenesisConfig

InherentDataProvider
Provides the slot duration inherent data for Aura.

Module

StakingSlasher
A type for performing slashing based on Aura reports.

Enums

Call

Constants

INHERENT_IDENTIFIER
The Aura inherent identifier.

Traits

AuraInherentData
Auxiliary trait to extract Aura inherent data.

HandleReport
Something that can handle Aura consensus reports.

Trait

Type Definitions

InherentType
The type of the Aura inherent.